Add missing auth mixins to GrantCreateView
The test_func method was defined, but without the UserPassesTestMixin,
Djano wasn't utilizing it.

As a result, any user could create a grant for any project, by using the
proper url (with the project primary key in the appropriate place).

Additionally, the LoginRequiredMixin was absent, allowing anyone to make
such a change.