Add prominent note that the integration test doesn't mock anything
The mock construction in the other test is a bit buried, so adding this
note clarifies to the reader immediately, rather than having them
perhaps believe all test cases use the mocks.

Not sure how to organize these tests, but having the integration test in
a separate class - but using the same or copy/paste'd data - didn't
really feel right either.