Add utility mixin for Views, to use snake_case in a view's template name
Presently, this only makes sense for subclasses of CreateView, although
it could be modified in the future to adapt with other View classes.

The mixin requires both the model and the template_name_suffix to be
specified as class attributes. (That is normally required of a
CreateView anyways.)

ColdFront currently uses snake_case ubiquitously in paths, so
CreateView's default of using simply lowercase (non-underscored) would
be a style deviation and probably cause continual confusion/frustration
for anyone to work on ResearchOutput templates in the future.

This implemented as a mixin reduces the friction of having another app
with multiple words in the future, being robust to most foreseeable
model names.

For this app, implementation as a mixin cleans up the View code,
meanwhile making the other "clean" View implementation option
(hardcoding) unattractive.