Standardize icons for Publications, Research Outputs (fa-newspaper use)
Publications previously was using both:
- 'fas fa-newspaper'
- 'far fa-newspaper'

Considering 'fas' was being used for the Publications listing (which
would be seen by all project members and reviewers) whereas 'far' just
at the ~minibar at the top (which would be seen only be those able to
edit the project), 'fas fa-newspaper' was probably most intended, and
would be most recognizeable going forward.

This frees up 'far fa-newspaper' to be used for Research Outputs.