Add a basic display of Research Outputs stats in Center Summary
This was not originally scoped in the Research Outputs work, but a quick
grep for the font awesome icons (to ensure completeness) pointed it out.

As such, a minimal stub (with real info) has been added to the page, so
that it won't go fully neglected, especially to stakeholders who use
that page.

However, since the Center Summary was outside the present scope of work,
any improvements can be done as future work.