Allow all staff (not just superusers) to access the User Search feature
For completeness, a "new" navbar menu has been created: "Staff", for
nonadmin staff.

The admin's navbar subtemplate has been copied to a new subtemplate:
navbar_nonadmin_staff.html

Therein:
- /admin has been removed (although is_staff does mean that URL is
  accessible within Django), and
- User Search is always available, but
- all other items have a permissions check before being listed.

Note that this navbar strategy does not appear to scale well - in fact,
navbar_director.html may need to be updated similarly to this new
subtemplate, as it contains items for which the supertemplate did not
check permissions on.