Extend UserProfileView/template to enable viewing other users' profiles
This ability requires being superuser/staff.

Because it doesn't make sense for superuser/staff to utilize the
"Upgrade Account" button here for other user profiles, that is only
displayed for one's own profile (if appropriate).

Username was selected for the url part, vs alternatives, because:
- vs id:          it's human-readable/rememberable, and it doesn't
                  "leak" pk data
- vs first/last:  it's nonoptional, generally shorter, and generally
                  typeable ASCII

Additionally, username is now displayed in a few parts of the template.
This should make it easier to work with multiple User Profile pages
open.

The url isn't integrated anywhere else in the app yet - forthcoming.