Add new feature: User Projects and Managers listing
The main intent of the feature is to illustrate who is responsible for a
given user. It also makes it easy to view which projects a user is part
of, with a subset of detail.

Because of the latter, it has some use to all users.

Viewing the listing for another username requires being superuser/staff.
The listing is accessible by an ordinary user only for their own
username.

To remove redundant info, the table for managers:
- is called 'Additional Managers'
- does not display PI info (it's displayed right above the table)
- does not display a line for the user, if the user is Manager (instead
  displaying this at the top-right of the table heading)

Additionally, the projects listed are sorted as follows (highest
priority first, placing at top):
- whether the user is the PI of the project
- whether the user is a manager of the project
- <less important details to note here, see queryset order_by>

The sorting above is intended to draw attention to projects that a user
is responsible for vs. projects that a user is simply part of.

As with the UserProfile url, the username is the key url part.

The url isn't integrated anywhere else in the app yet - forthcoming.