Fix year validation bug.
This fixes #258. Add better form validation for manual publication
entry. Change to using uuid4 instead of sha256 hash for unique ids. The
issue is when manually entering a publication, we do not ask for a DOI
as it may not have one. As a doi uniquely identifies a publication, we
simply use a randomly generated uuid4 as the unique id.