Skip allocations to inactive resources.
This fixes #268. When adding users to a unix group, the freeipa plugin checks a
users active allocations (where they are an active user on said allocation).
This commit adds a check to see if all resources for the allocation are set to
inactive. If so, the user will not be added to the unix group.