First draft of Attribute Expansion project
We allow for a new AttributeType "Attribute Expanded Text"
This is basically just a Text field which can have parameters (typically
other attributes or simple calculations of attributes) can be expanded.

The process works as follows:
Let testalloc be an Allocation with testresource as a Resource, and
having the Attribute of type "SU" with value of 10000.

We want allocations of testresource to get cpu limitted by the number of SUs in
Slurm, so the Resource testresource defines the ResourceAttribute
"slurm_specs" with AttributeType "Attribute Expanded Text" and
with the value "GrpTRESMins=cpu={cpumin}"

It also defines a normal Text attribute of ResourceAttributeType
"slurm_specs_attriblist" with the value:
cpumin := SU
cpumin |= 0
cpumin *= 60

The various additions cause slurm_specs_attriblist to create a temporary
"apdict" with key cpumin set to the value of the Attribute named "SU"
related to the allocation.  The next line in the attriblist will default
the value of cpumin to 0 if SU was not found.  The last line multiplies
it by 60 to convert from CPU-hours to CPU-min.

This apdict is then used as the argument of a python format() function
used to expand the slurm_specs attribute, so when the slurm plugin
writes out the allocation, it will look something like
Account - 'testalloc':GrpTRESMin=cpu=600000

This is achieved by:
* adding attribute_expansion.py under core, this defines a number of
routines that do most of the work
* adding the method typed_value() to AllocationAttribute and
ResourceAttribute: this basically converts the value of the attribute to
an int or float if the base AttributeType is Int or Float.  This is
needed to get proper arithmetic when expanding parameters.
* adding the method expanded_value() to AllocationAttribute and
ResourceAttribute: this returns the expanded value as described above.
It takes an optional parameter 'typed' (default true) which determines
whether we return the value as the proper python type (a la
typed_value), and extra_allocations (a list of Allocation, defaulting to
the empty list), giving some extra Allocations whose attributes will be
considered when searching for Attributes in the attriblist (this is so
slurm_specs can be defined in the Resource and SU in the Allocation)
* Adding some extra optional parameters to get_attribute() and
get_attribute_list() methods of Allocation and Resource.  The extra
parameters are:
    - expand: boolean, default True.  Whether to 'expand' any values for
	attributes of 'Attribute Expanded Text' in the returned list.
    - typed: boolean, default True.  Whether to 'type' any values for
	attributes of 'Attribute Expanded Text' in the returned list.
    - extra_allocations: list of Allocations, default [].  Extra
	allocations we can search for attributes in

Finally a small change to SlurmCluster.new_from_resource to pass the
child allocation via the extra_allocations list to get_attribute_list
when compute partion_specs and partition_user_specs.