Add example of 'Attribute Expanded Text' to test data
Create a 'University Metered HPC' cluster, with slurm_specs
and slurm_specs_attriblist set to limit usage of allocations
of this resource based on the value of the allocations Core Usage
attribute.

Also create an allocation for the metered HPC for PI cgray with
Core Usage set to 150,000.

slurm_dump should produce an 'Account' record limiting usage to
60 * 150,000 CPU minutes.