Updates to support TLS in ldap_user_search plugin
This PR should address concerns of #631

Basically:
1) Adds new config parameter LDAP_USER_SEARCH_CERT_VALIDATE_MODE
which is passed to the ldap3.Tls constructor as validate.
Accepts as values:
	'required' : Certs are required and must validate
	'optional' : Certs are optional, but must validate if provided
	'none' (or None): Certs are ignored.
The default is None

2) The LDAP_USER_SEARCH_CERT_VALIDATE_MODE is passed as the validate
field to the ldap3.Tls constructor

3) If LDAP_USE_TLS is set, we pass the connection parameter 'auto_bind'
as ldap3.AUTO_BIND_TLS_BEFORE_BIND instead of simply True
Inspection of ldap3 code shows that when this parameter is set
to True (a value which is no longer listed in docs as valid) it
is treated as AUTO_BIND_NO_TLS, so the previous before of leaving
this as True was not doing TLS despite claiming to do TLS.
This fix should change that.